chore: restructure model classes and functions
* move enums to `enums.py` module
* move models to `models.py` module
* explicitly re-export all "public" functions and classes in `__init__.py`
* adapt readme
