+++ src/bdew_datetimes/german_time_zone.py
@@ -1,0 +1,5 @@
+"""static timezone object for Berlin/Germany"""
+
+from pytz import timezone
+
+GERMAN_TIME_ZONE = timezone("Europe/Berlin")
